% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinGltf.R
\name{checkFilesBinGltf}
\alias{checkFilesBinGltf}
\title{Check .bin and .gltf Files}
\usage{
checkFilesBinGltf(xmlPath, nlods)
}
\arguments{
\item{nlods}{}
}
\value{

}
\description{
Check .bin and .gltf Files
}
