% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinGltf.R
\name{fixBinGltf}
\alias{fixBinGltf}
\title{Fix missing bin/gltf}
\usage{
fixBinGltf(PackageSourcesDir, invalids, deleteTextures = TRUE)
}
\arguments{
\item{PackageSourcesDir}{Path to PackageSources directory.}

\item{invalids}{List with invalid files from `checkAllFilesBinGltf`.}

\item{deleteTextures}{Whether to delete textures in modelLib/texture related to the corrupted data.}
}
\description{
Fix missing bin/gltf
}
