% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{fixLods}
\alias{fixLods}
\title{Fix Lods}
\usage{
fixLods(PackageSourcesDir, invalids, deleteTextures = TRUE)
}
\arguments{
\item{PackageSourcesDir}{Path to PackageSources directory.}

\item{invalids}{List with invalid files from `checkAllLods`.}

\item{deleteTextures}{Whether to delete textures in modelLib/texture related to the corrupted data.}
}
\description{
Fix Lods
}
