% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSceneryFiles.R
\name{getXmlGuid}
\alias{getXmlGuid}
\title{Get GUID from .xml}
\usage{
getXmlGuid(xmlPath)
}
\arguments{
\item{xmlPath}{Path to .xml in modelLib.}
}
\value{
GUID related to the .xml provided.
}
\description{
Get GUID from a modelLib .xml file.
}
\examples{
\donttest{
getXmlGuid("D:/FSProjects/my-package/PackageSources/modelLib/03726152506153634.xml")
}
}
