% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCorruptedFiles.R
\name{isValidXML}
\alias{isValidXML}
\title{Check if a object is valid}
\usage{
isValidXML(xmlPath, nlods)
}
\arguments{
\item{xmlPath}{Path of the xml file.}
}
\value{
Result of validation for both lods corrupted and missing bin or gltf files.
}
\description{
Check if a object is valid
}
