% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectsXml.R
\name{objectsXmlGuids}
\alias{objectsXmlGuids}
\title{Get guids from objects.xml nodes}
\usage{
objectsXmlGuids(xmlPath)
}
\arguments{
\item{xmlPath}{}
}
\description{
Get guids from objects.xml nodes
}
\examples{
objectsXmlNodes("D:/FSProjects/my-package/PackageSources/scene/objects.xml")
}
