% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{removeBinGltfByLods}
\alias{removeBinGltfByLods}
\title{Remove .bin and .gltf by lods}
\usage{
removeBinGltfByLods(modelLibDir, lodsToRemove)
}
\arguments{
\item{modelLibDir}{Path to modelLib directory.}

\item{lodsToRemove}{A (character) vector with lods to remove.
Example: if you wan't to remove LOD01 and LOD02, provide a vector c("01", "02").}
}
\description{
Remove .bin and .gltf by lods
}
