% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{removeLodNodesFromXML}
\alias{removeLodNodesFromXML}
\title{Remove lod nodes from XML}
\usage{
removeLodNodesFromXML(xmlPath, lodsToRemove)
}
\arguments{
\item{xmlPath}{Path of the xml file of the object.}

\item{lodsToRemove}{A (character) vector with lods to remove.
Example: if you wan't to remove LOD01 and LOD02, provide a vector c("01", "02").}
}
\description{
Remove lod nodes from XML
}
