% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{removeLods}
\alias{removeLods}
\title{Remove lods from project}
\usage{
removeLods(PackageSourcesDir, lodsToRemove, removeBinGltf = TRUE,
  removeTextures = TRUE)
}
\arguments{
\item{PackageSourcesDir}{Path to PackageSources directory.}

\item{lodsToRemove}{A (character) vector with lods to remove.
Example: if you wan't to remove LOD01 and LOD02, provide a vector c("01", "02").}

\item{removeBinGltf}{Logical. Whether to remove correspondent .bin and .gltf files.}

\item{removeTextures}{Logical. Whether to remove correspondent texture files.}
}
\description{
Remove lods from project
}
