% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{removeModelLibTexturesByLods}
\alias{removeModelLibTexturesByLods}
\title{Remove textures by lods}
\usage{
removeModelLibTexturesByLods(textureDir, lodsToRemove)
}
\arguments{
\item{textureDir}{Path to modelLib/texture directory.}

\item{lodsToRemove}{A (character) vector with lods to remove.
Example: if you wan't to remove LOD01 and LOD02, provide a vector c("01", "02").}
}
\description{
Remove textures by lods
}
