% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinGltf.R
\name{checkAllFilesBinGltf}
\alias{checkAllFilesBinGltf}
\title{Check multiple .bin and .gltf files}
\usage{
checkAllFilesBinGltf(modelLibDir, nlods)
}
\arguments{
\item{nlods}{Numeric. Number of lods of the project.}
}
\description{
Check multiple .bin and .gltf files
}
