% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCorruptedFiles.R
\name{fixCorruptedFiles}
\alias{fixCorruptedFiles}
\title{Detect and Remove Corrupted Files}
\usage{
fixCorruptedFiles(PackageSourcesDir, nlods)
}
\arguments{
\item{PackageSourcesDir}{}
}
\value{

}
\description{
Detect and Remove Corrupted Files
}
