% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectsXml.R
\name{fixObjectsXML}
\alias{fixObjectsXML}
\title{Fix objects.xml}
\usage{
fixObjectsXML(xmlPath, invalidGuids, createBackup = FALSE)
}
\arguments{
\item{xmlPath}{Path to the objects.xml file (usually located inside PackageSources/scene).}

\item{createBackup}{}
}
\description{
Fix objects.xml
}
