% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectsXml.R
\name{objectsXmlGuids}
\alias{objectsXmlGuids}
\title{Get guids from objects.xml nodes}
\usage{
objectsXmlGuids(xmlPath)
}
\arguments{
\item{xmlPath}{Path to the objects.xml file (usually located inside PackageSources/scene).}
}
\value{
A vector with the guids of all objects.xml nodes.
}
\description{
Get all guids identifiers registered in objects.xml file.
}
\examples{
\dontrun{
objectsXmlGuids("D:/FSProjects/my-package/PackageSources/scene/objects.xml")
}
}
