% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{removeLodNodesFromXML}
\alias{removeLodNodesFromXML}
\title{Remove lod nodes from XML}
\usage{
removeLodNodesFromXML(xmlPath, lodsToRemove)
}
\arguments{
\item{removeTextures}{}
}
\description{
Remove lod nodes from XML
}
