% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{removeLods}
\alias{removeLods}
\title{Remove lods from project}
\usage{
removeLods(PackageSourcesDir, lodsToRemove, removeBinGltf = TRUE,
  removeTextures = TRUE)
}
\arguments{
\item{lodsToRemove}{}
}
\description{
Remove lods from project
}
